test: fix unit tests based on changes
