+++ packages/core/admin/admin/src/core/store/configure.ts
@@ -4,24 +4,21 @@ import {
   Middleware,
   Reducer,
   combineReducers,
-  createSelector,
-  Selector,
 } from '@reduxjs/toolkit';
-import { useDispatch, useStore, TypedUseSelectorHook, useSelector } from 'react-redux';
 
-import { RBACReducer } from '../components/RBACProvider';
+import { RBACReducer } from '../../components/RBACProvider';
 // @ts-expect-error no types, yet.
-import rbacManagerReducer from '../content-manager/hooks/useSyncRbac/reducer';
+import rbacManagerReducer from '../../content-manager/hooks/useSyncRbac/reducer';
 // @ts-expect-error no types, yet.
-import cmAppReducer from '../content-manager/pages/App/reducer';
+import cmAppReducer from '../../content-manager/pages/App/reducer';
 // @ts-expect-error no types, yet.
-import editViewLayoutManagerReducer from '../content-manager/pages/EditViewLayoutManager/reducer';
+import editViewLayoutManagerReducer from '../../content-manager/pages/EditViewLayoutManager/reducer';
 // @ts-expect-error no types, yet.
-import listViewReducer from '../content-manager/pages/ListView/reducer';
+import listViewReducer from '../../content-manager/pages/ListView/reducer';
 // @ts-expect-error no types, yet.
-import editViewCrudReducer from '../content-manager/sharedReducers/crudReducer/reducer';
+import editViewCrudReducer from '../../content-manager/sharedReducers/crudReducer/reducer';
 // @ts-expect-error no types, yet.
-import appReducer from '../pages/App/reducer';
+import appReducer from '../../pages/App/reducer';
 
 /**
  * @description Static reducers are ones we know, they live in the admin package.
@@ -89,20 +86,6 @@ type Store = ReturnType<typeof configureStoreImpl> & {
 };
 
 type RootState = ReturnType<Store['getState']>;
-type AppDispatch = Store['dispatch'];
 
-const useTypedDispatch: () => AppDispatch = useDispatch;
-const useTypedStore = useStore as () => Store;
-const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
-
-const createTypedSelector = <TResult>(selector: Selector<RootState, TResult>) =>
-  createSelector((state: RootState) => state, selector);
-
-export {
-  useTypedDispatch,
-  useTypedStore,
-  useTypedSelector,
-  configureStoreImpl as configureStore,
-  createTypedSelector,
-};
+export { configureStoreImpl as configureStore };
 export type { RootState, Store };
