+++ packages/core/admin/admin/src/hooks/useInjectReducer/tests/useInjectReducer.test.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { renderHook } from '@testing-library/react';
 import { Provider } from 'react-redux';
 
-import { configureStore } from '../../../core/store';
+import { configureStore } from '../../../core/store/configure';
 import { useInjectReducer } from '../useInjectReducer';
 
 const store = configureStore();
